replaced all douubles with floats

+++ include/hgwosca.h
@@ -7,20 +7,20 @@ extern "C" {
 #include "stddef.h"
 #include "utils.h"
 
-double * gwo_hgwosca(obj_func_t obj_func,
+float * gwo_hgwosca(obj_func_t obj_func,
                      size_t wolf_count,
                      size_t dim,
                      size_t max_iterations,
-                     double min_position,
-                     double max_position);
+                     float min_position,
+                     float max_position);
 
 /**
    Initialise population of `wolf_count` wolves, each with `dim` dimensions, where
    each dimension is bound by `min_positions` and `max_positions`.
  */
-void gwo_init_population(double *population,
+void gwo_init_population(float *population,
                          size_t wolf_count, size_t dim,
-                         double min_position, double max_position);
+                         float min_position, float max_position);
 
 /**
    Update the fitness of all wolves.
@@ -28,33 +28,33 @@ void gwo_init_population(double *population,
 void gwo_update_fitness(size_t wolf_count,
                         size_t dim,
                         obj_func_t obj_func,
-                        double *population,
-                        double *fitness);
+                        float *population,
+                        float *fitness);
 
 /**
    Find the new wolf leaders.
  */
-void gwo_update_leaders(size_t wolf_count, double *fitness, size_t *alpha, size_t *beta, size_t *delta);
+void gwo_update_leaders(size_t wolf_count, float *fitness, size_t *alpha, size_t *beta, size_t *delta);
 
 /**
    Initialise the fitness for all wolves.
  */
-void gwo_init_fitness(double *fitness, size_t wolf_count,
-                      size_t dim, obj_func_t obj_func, double *population);
+void gwo_init_fitness(float *fitness, size_t wolf_count,
+                      size_t dim, obj_func_t obj_func, float *population);
 
 /**
    Get recommended position of `dimension` for `wolf` with respect to `leader`, for a given
    value of `a`.
    TODO: test
  */
-double gwo_get_wolf_pos_update_dim_leader(size_t dimension, double a, const double *wolf, const double *leader_pos);
+float gwo_get_wolf_pos_update_dim_leader(size_t dimension, float a, const float *wolf, const float *leader_pos);
 
 /**
    Get recommended position of `dimension` for `wolf` with respect to the alpha, for a given
    value of `a`. See equ (12) on the Hybrid paper.
    TODO: test
  */
-double gwo_get_wolf_pos_update_dim_alpha(size_t dimension, double a, const double *wolf, const double *alpha_pos);
+float gwo_get_wolf_pos_update_dim_alpha(size_t dimension, float a, const float *wolf, const float *alpha_pos);
 
 
 /**
@@ -63,11 +63,11 @@ double gwo_get_wolf_pos_update_dim_alpha(size_t dimension, double a, const doubl
    TODO: test
  */
 void gwo_update_wolf_position(size_t dim,
-                              double a,
-                              double *wolf,
-                              const double *alpha_pos,
-                              const double *beta_pos,
-                              const double *delta_pos);
+                              float a,
+                              float *wolf,
+                              const float *alpha_pos,
+                              const float *beta_pos,
+                              const float *delta_pos);
 
 /**
    Update the positions of all wolves.
@@ -75,23 +75,23 @@ void gwo_update_wolf_position(size_t dim,
  */
 void gwo_update_all_positions(size_t wolf_count,
                               size_t dim,
-                              double a,
-                              double *population,
+                              float a,
+                              float *population,
                               size_t alpha, size_t beta, size_t delta);
 
 
 /**
    Clamp `val` between `min` and `max`.
  */
-double gwo_clamp(double val, double min, double max);
+float gwo_clamp(float val, float min, float max);
 
 /**
    Clamp solutions into feasible space.
  */
-void gwo_clamp_all_positions(size_t wolf_count, size_t dim, double *population,
-                             double min_position, double max_position);
+void gwo_clamp_all_positions(size_t wolf_count, size_t dim, float *population,
+                             float min_position, float max_position);
 
-size_t gwo_get_fittest_idx(size_t colony_size, const double *fitness);
+size_t gwo_get_fittest_idx(size_t colony_size, const float *fitness);
 
 #ifdef __cplusplus
 }
